requirements.txt
streamlit>=1.36
gspread>=6.0.0
google-auth>=2.30.0
pandas>=2.2.2
